Extract PearbotCommand base class into separate file
Split this out to avoid duplication, also clean up indentation and excess whitespace
