+++ lib/pearbot/shared_commands.rb
@@ -1,59 +1,46 @@
 module Pearbot
-    module SharedCommands
-      class PearbotCommand < SlackRubyBot::Commands::Base
-        def self.replace_me_with_id(parsed_id, current_user_id)
-          parsed_id == "me" ? current_user_id : parsed_id
-        end
+  module SharedCommands
+    class Snooze < PearbotCommand
+      match /snooze ?(me)/
 
-        def self.format_date_time(timestamp)
-          unix = timestamp.to_i
-          fallback = timestamp.strftime("%A, %B #{timestamp.day.ordinalize}, %Y at %H.%M UTC")
-
-          "<!date^#{unix}^{date_long_pretty} at {time}|#{fallback}>"
-        end
-      end
-      class Snooze < PearbotCommand
-        match /snooze ?(me)/
-
-        help do
-          title 'snooze me'
-          desc 'Temporarily disable drawing for yourself.'
-        end
-
-        def self.call(client, data, match)
-          participant = Participant.find_by(slack_user_id: data.user)
-          pool = participant.pools.last
-          pool.refresh_participants if pool.present?
-
-          if pool.blank?
-            client.say(channel: data.channel, text: "🙅‍♀️ No pool exists ", gif: 'no')
-          else
-            participant.snooze_pool(pool)
-            client.say(channel: data.channel, text: "Snoozed drawing for #{participant.name} in <##{pool.slack_channel_id}>. 😴", gif: 'sleep')
-          end
-        end
+      help do
+        title 'snooze me'
+        desc 'Temporarily disable drawing for yourself.'
       end
 
-      class Resume < PearbotCommand
-        match /resume ?(me)/
+      def self.call(client, data, match)
+        participant = Participant.find_by(slack_user_id: data.user)
+        pool = participant.pools.last
+        pool.refresh_participants if pool.present?
 
-        help do
-          title 'resume me/@user'
-          desc 'Re-enables drawing for yourself.'
+        if pool.blank?
+          client.say(channel: data.channel, text: "🙅‍♀️ No pool exists ", gif: 'no')
+        else
+          participant.snooze_pool(pool)
+          client.say(channel: data.channel, text: "Snoozed drawing for #{participant.name} in <##{pool.slack_channel_id}>. 😴", gif: 'sleep')
         end
+      end
+    end
+    class Resume < PearbotCommand
+      match /resume ?(me)/
 
-        def self.call(client, data, match)
-          participant = Participant.find_by(slack_user_id: data.user)
-          pool = participant.pools.last
-          pool.refresh_participants if pool.present?
+      help do
+        title 'resume me/@user'
+        desc 'Re-enables drawing for yourself.'
+      end
 
-          if pool.blank?
-            client.say(channel: data.channel, text: "🙅‍♀️ No pool exists ", gif: 'no')
-          else
-            participant.resume_pool(pool)
-            client.say(channel: data.channel, text: "Resumed drawing for #{participant.name} in <##{pool.slack_channel_id}>. 😊", gif: 'awake')
-          end
+      def self.call(client, data, match)
+        participant = Participant.find_by(slack_user_id: data.user)
+        pool = participant.pools.last
+        pool.refresh_participants if pool.present?
+
+        if pool.blank?
+          client.say(channel: data.channel, text: "🙅‍♀️ No pool exists ", gif: 'no')
+        else
+          participant.resume_pool(pool)
+          client.say(channel: data.channel, text: "Resumed drawing for #{participant.name} in <##{pool.slack_channel_id}>. 😊", gif: 'awake')
         end
       end
     end
   end
+end
